+++ src/Process/TearDownProcess.php
@@ -1,0 +1,9 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Steevanb\ParallelProcess\Process;
+
+class TearDownProcess extends Process implements TearDownProcessInterface
+{
+}