+++ tests/Console/Application/Theme/DefaultTheme/OutputStartTest.php
@@ -7,7 +7,7 @@ namespace Steevanb\ParallelProcess\Tests\Console\Application\Theme\DefaultTheme;
 use PHPUnit\Framework\TestCase;
 use Steevanb\ParallelProcess\{
     Console\Application\Theme\DefaultTheme,
-    Process\ProcessArray,
+    Process\ProcessInterfaceArray,
     Tests\Console\Output\TestOutput,
     Tests\CreateLsProcessTrait
 };
@@ -22,7 +22,7 @@ final class OutputStartTest extends TestCase
         $output = new TestOutput();
         (new DefaultTheme())->outputStart(
             $output,
-            new ProcessArray()
+            new ProcessInterfaceArray()
         );
 
         static::assertSame('', $output->getOutputed());
@@ -33,7 +33,7 @@ final class OutputStartTest extends TestCase
         $output = (new TestOutput())->setDecorated(true);
         (new DefaultTheme())->outputStart(
             $output,
-            new ProcessArray()
+            new ProcessInterfaceArray()
         );
 
         static::assertSame('', $output->getOutputed());
@@ -44,7 +44,7 @@ final class OutputStartTest extends TestCase
         $output = new TestOutput();
         (new DefaultTheme())->outputStart(
             $output,
-            new ProcessArray([$this->createLsProcess()])
+            new ProcessInterfaceArray([$this->createLsProcess()])
         );
 
         static::assertSame(
@@ -58,7 +58,7 @@ final class OutputStartTest extends TestCase
         $output = (new TestOutput())->setDecorated(true);
         (new DefaultTheme())->outputStart(
             $output,
-            new ProcessArray([$this->createLsProcess()])
+            new ProcessInterfaceArray([$this->createLsProcess()])
         );
 
         static::assertSame(
@@ -75,7 +75,7 @@ final class OutputStartTest extends TestCase
         $process2 = $this->createLsProcess();
         $process2->mustRun();
 
-        $processes = new ProcessArray([$process1, $process2]);
+        $processes = new ProcessInterfaceArray([$process1, $process2]);
 
         $output = new TestOutput();
 
@@ -95,7 +95,7 @@ final class OutputStartTest extends TestCase
         $process2 = $this->createLsProcess();
         $process2->mustRun();
 
-        $processes = new ProcessArray([$process1, $process2]);
+        $processes = new ProcessInterfaceArray([$process1, $process2]);
 
         $output = (new TestOutput())->setDecorated(true);