+++ tests/Process/ProcessInterfaceArray/ConstructTest.php
@@ -2,12 +2,12 @@
 
 declare(strict_types=1);
 
-namespace Steevanb\ParallelProcess\Tests\Process\ProcessArray;
+namespace Steevanb\ParallelProcess\Tests\Process\ProcessInterfaceArray;
 
 use PHPUnit\Framework\TestCase;
 use Steevanb\ParallelProcess\{
-    Process\Process,
-    Process\ProcessArray
+    Process\ProcessInterface,
+    Process\ProcessInterfaceArray
 };
 use Steevanb\PhpTypedArray\{
     NullValueModeEnum,
@@ -15,15 +15,15 @@ use Steevanb\PhpTypedArray\{
     ValueAlreadyExistsModeEnum
 };
 
-/** @covers \Steevanb\ParallelProcess\Process\ProcessArray::__construct */
+/** @covers \Steevanb\ParallelProcess\Process\ProcessInterfaceArray::__construct */
 final class ConstructTest extends TestCase
 {
     public function testDefaultValues(): void
     {
-        $processes = new ProcessArray();
+        $processes = new ProcessInterfaceArray();
 
         static::assertCount(0, $processes);
-        static::assertSame(Process::class, $processes->getClassName());
+        static::assertSame(ProcessInterface::class, $processes->getClassName());
         static::assertSame(ObjectComparisonModeEnum::STRING, $processes->getComparisonMode());
         static::assertSame(ValueAlreadyExistsModeEnum::ADD, $processes->getValueAlreadyExistMode());
         static::assertSame(NullValueModeEnum::ALLOW, $processes->getNullValueMode());